build: refactor scripts directory to standard structure
