+++ scripts/discovery/discover_examples.py
